Rework README.md, add key generation, fix host file
Most of what I had documented in the README.md is in the
bootstrap.sh file. So I changed the docs to reference it but still
listed the steps.

Added the ssh-keygen command to the bootstrap.sh file.

Added the cephrgw container to the cephadmin's etc/hosts file

Removed the --override-conf option that was suggested in the
ansible install error message

Added the "echo "done" to the file shell script command in the
install_rgw file.